requirements.txt
requests~=2.31.0
hexbytes~=0.3.1
loguru~=0.7.2
web3~=6.18.0
urllib3~=2.2.1
art~=6.2
PyYAML~=6.0.1